#179 ➕ `iolanta.parse_quads`
